llmgs/pyproject.toml
[project]
name = "llmgs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "orjson>=3.10.16",
    "pydantic-ai-slim[mcp]>=0.1.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "ujson>=5.10.0",
]
